refactor dbConnection to allow importing from everywhere
