+++ packages/cli/src/index.ts
@@ -1,4 +1,5 @@
-export { execute, dbConnectAsync } from './lib/setup';
+export { execute } from './lib/setup';
+export { dbConnectAsync } from './lib/setup/dbConnection';
 // required by main.ts
 export { Vendor } from './lib/setup/setupVendor';
 // required by main.ts
